Update README.md
In the home page of this recipe focused website, you can see a few buttons like home, quick links, chicken, red meat, pasta, reviews, contact, and a quick panel. If you click on the quick panel, you get a link to a pinterest recipe page. The home button has a quick mention of the website and how food is a multicultural language. The quick links was initially meant to give access to multiple different kinds of recipes but it ended up blank. The chicken button gives a link to 69 different chicken recipes. The red meat button give a quick link to 40 easy recipes with different kinds of beef. The pasta button also give a link to 47 pasta recipes. The reviews button give two other options, submit review, and view reviews. If you click on submit review a form will appear asking personal information and a review text input. The view reviews page was never worked on so it ended up coming out blank. Lastly, the contact button shows a small grid with the contact information on how to get in touch with the creators of the website.